requirements.txt
langchain-community
sentence-transformers
langchain
fastapi
uvicorn
pydantic
faiss-cpu
google-generativeai
numpy